requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile requirements.in
#
--index-url https://mirrors.aliyun.com/pypi/simple/
--trusted-host mirrors.aliyun.com

annotated-types==0.7.0
    # via pydantic
anyio==4.10.0
    # via
    #   httpx
    #   starlette
    #   watchfiles
backports-asyncio-runner==1.2.0
    # via pytest-asyncio
build==1.3.0
    # via pip-tools
certifi==2025.8.3
    # via
    #   httpcore
    #   httpx
    #   sentry-sdk
click==8.2.1
    # via
    #   pip-tools
    #   rich-toolkit
    #   typer
    #   uvicorn
coverage[toml]==7.10.4
    # via pytest-cov
dnspython==2.7.0
    # via email-validator
email-validator==2.2.0
    # via
    #   fastapi
    #   pydantic
exceptiongroup==1.3.0
    # via
    #   anyio
    #   pytest
fastapi[all]==0.116.1
    # via -r requirements.in
fastapi-cli[standard]==0.0.8
    # via fastapi
fastapi-cloud-cli==0.1.5
    # via fastapi-cli
h11==0.16.0
    # via
    #   httpcore
    #   uvicorn
httpcore==1.0.9
    # via httpx
httptools==0.6.4
    # via uvicorn
httpx==0.28.1
    # via
    #   fastapi
    #   fastapi-cloud-cli
idna==3.10
    # via
    #   anyio
    #   email-validator
    #   httpx
iniconfig==2.1.0
    # via pytest
itsdangerous==2.2.0
    # via fastapi
jinja2==3.1.6
    # via fastapi
markdown-it-py==4.0.0
    # via rich
markupsafe==3.0.2
    # via jinja2
mdurl==0.1.2
    # via markdown-it-py
orjson==3.11.2
    # via fastapi
packaging==25.0
    # via
    #   build
    #   pytest
pip-tools==7.5.0
    # via -r requirements.in
pluggy==1.6.0
    # via
    #   pytest
    #   pytest-cov
pydantic[email]==2.11.7
    # via
    #   fastapi
    #   fastapi-cloud-cli
    #   pydantic-extra-types
    #   pydantic-settings
pydantic-core==2.33.2
    # via pydantic
pydantic-extra-types==2.10.5
    # via fastapi
pydantic-settings==2.10.1
    # via fastapi
pygments==2.19.2
    # via
    #   pytest
    #   rich
pyproject-hooks==1.2.0
    # via
    #   build
    #   pip-tools
pytest==8.4.1
    # via
    #   -r requirements.in
    #   pytest-asyncio
    #   pytest-cov
    #   pytest-mock
pytest-asyncio==1.1.0
    # via -r requirements.in
pytest-cov==6.2.1
    # via -r requirements.in
pytest-mock==3.14.1
    # via -r requirements.in
python-dotenv==1.1.1
    # via
    #   pydantic-settings
    #   uvicorn
python-multipart==0.0.20
    # via fastapi
pyyaml==6.0.2
    # via
    #   fastapi
    #   uvicorn
rich==14.1.0
    # via
    #   rich-toolkit
    #   typer
rich-toolkit==0.15.0
    # via
    #   fastapi-cli
    #   fastapi-cloud-cli
rignore==0.6.4
    # via fastapi-cloud-cli
sentry-sdk==2.35.0
    # via fastapi-cloud-cli
shellingham==1.5.4
    # via typer
sniffio==1.3.1
    # via anyio
starlette==0.47.2
    # via fastapi
tomli==2.2.1
    # via
    #   build
    #   coverage
    #   pip-tools
    #   pytest
typer==0.16.1
    # via
    #   fastapi-cli
    #   fastapi-cloud-cli
typing-extensions==4.14.1
    # via
    #   anyio
    #   exceptiongroup
    #   fastapi
    #   pydantic
    #   pydantic-core
    #   pydantic-extra-types
    #   rich-toolkit
    #   starlette
    #   typer
    #   typing-inspection
    #   uvicorn
typing-inspection==0.4.1
    # via
    #   pydantic
    #   pydantic-settings
ujson==5.10.0
    # via fastapi
urllib3==2.5.0
    # via sentry-sdk
uvicorn[standard]==0.35.0
    # via
    #   fastapi
    #   fastapi-cli
    #   fastapi-cloud-cli
uvloop==0.21.0
    # via uvicorn
watchfiles==1.1.0
    # via uvicorn
websockets==15.0.1
    # via uvicorn
wheel==0.45.1
    # via pip-tools

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
